tests/submission/Microsoft/video/video_021.htm

<!DOCTYPE HTML>
<html>
<head>
  <title>HTML5 Media Elements: 'preload' IDL attribute reflects the content attribute. ("metadata")</title>
  <meta content="text/html; charset=UTF-8" http-equiv="Content-Type">
  <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
  <link rel="help" href="http://www.w3.org/TR/html5/video.html#attr-media-preload" />
  <meta name="assert" content="'preload' IDL attribute reflects the content attribute. ("metadata")" />  
</head>
<body>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <div id=log></div>
  <video id="video0" preload="metadata">Your browser does not support media elements.</video>
  <script type="text/javascript">
    test(function() { assert_equals(document.getElementById("video0").preload, "metadata")}, "'preload' IDL attribute reflects the content attribute.");
  </script>
</div>
</body>
</html>
